Fixed organizationIdHash Bean creation
CHANGE:
Now the organizationID is in a new class called ApplicationConfig instead of DLTConfig
